+++ src/types/MatrixDense/MatrixDense_sgl.cu
@@ -232,10 +232,9 @@ MatrixDense<__half> MatrixDense<float>::to_half() const {
     
     MatrixDense<__half> created_mat(cu_handles, m_rows, n_cols);
 
-    double NUM_THREADS = 1024; // threads per thread block just 1 warp
-    double NUM_BLOCKS = static_cast<double>(
-        std::ceil(static_cast<double>(m_rows*n_cols)/static_cast<double>(NUM_THREADS))
-    );
+    int NUM_THREADS = 1024; // threads per thread block just maximum
+    int NUM_BLOCKS = std::ceil(static_cast<double>(m_rows*n_cols)/static_cast<double>(NUM_THREADS));
+
     generalmatrix_sgl_kernels::cast_to_half<<<NUM_THREADS, NUM_BLOCKS>>>(
         d_mat, created_mat.d_mat, m_rows*n_cols
     );
@@ -252,10 +251,9 @@ MatrixDense<double> MatrixDense<float>::to_double() const {
     
     MatrixDense<double> created_mat(cu_handles, m_rows, n_cols);
 
-    double NUM_THREADS = 1024; // threads per thread block just 1 warp
-    double NUM_BLOCKS = static_cast<double>(
-        std::ceil(static_cast<double>(m_rows*n_cols)/static_cast<double>(NUM_THREADS))
-    );
+    int NUM_THREADS = 1024; // threads per thread block just maximum
+    int NUM_BLOCKS = std::ceil(static_cast<double>(m_rows*n_cols)/static_cast<double>(NUM_THREADS));
+
     generalmatrix_sgl_kernels::cast_to_double<<<NUM_THREADS, NUM_BLOCKS>>>(
         d_mat, created_mat.d_mat, m_rows*n_cols
     );
